requirements.txt
astroid==2.3.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
Django==2.2
django-pwa==1.0.5
django-sslserver==0.21
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
oauthlib==3.1.0
Pillow==6.2.1
psycopg2==2.8.3
PyJWT==1.7.1
pylint==2.4.2
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
